Change GetHashCode calculation and update tests
