+++ Tests/FO-DICOM.Tests/DicomDatasetTest.cs
@@ -886,122 +886,6 @@ namespace FellowOakDicom.Tests
             Assert.False(dataset.TryGetSingleValue(DicomTag.SeriesNumber, out int _));
         }
 
-        [Fact]
-        public void Equals_SameObject_ShouldReturnTrue()
-        {
-            var dataset = new DicomDataset();
-            Assert.True(dataset.Equals(dataset));
-        }
-
-        [Fact]
-        public void Equals_NullObject_ShouldReturnFalse()
-        {
-            var dataset = new DicomDataset();
-            Assert.False(dataset.Equals(null));
-        }
-
-        [Fact]
-        public void Equals_DifferentType_ShouldReturnFalse()
-        {
-            var dataset = new DicomDataset();
-            var otherObject = new object();
-            Assert.False(dataset.Equals(otherObject));
-        }
-
-        [Fact]
-        public void Equals_DifferentEmptyDatasets_ShouldReturnTrue()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            Assert.True(dataset1.Equals(dataset2));
-        }
-
-        [Fact]
-        public void Equals_SameDataDatasets_ShouldReturnTrue()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            var tag = DicomTag.URNCodeValue;
-            dataset1.Add<string>(tag);
-            dataset2.Add<string>(tag);
-
-            Assert.True(dataset1.Equals(dataset2));
-        }
-
-        [Fact]
-        public void Equals_SameDataDatasetsWithDifferentAddOrder_ShouldReturnTrue()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            var tag1 = DicomTag.URNCodeValue;
-            var tag2 = DicomTag.SelectorLTValue;
-            dataset1.Add<string>(tag1);
-            dataset1.Add<string>(tag2);
-
-            dataset2.Add<string>(tag2);
-            dataset2.Add<string>(tag1);
-
-            Assert.True(dataset1.Equals(dataset2));
-        }
-
-        [Fact]
-        public void EqualityOperator_DifferentEmptyDatasets_ShouldReturnTrue()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            Assert.True(dataset1.Equals(dataset2));
-        }
-
-        [Fact]
-        public void EqualityOperator_SameDataDatasets_ShouldReturnTrue()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            var tag = DicomTag.URNCodeValue;
-            dataset1.Add<string>(tag);
-            dataset2.Add<string>(tag);
-
-            Assert.True(dataset1 == dataset2);
-        }
-
-        [Fact]
-        public void EqualityOperator_DifferentDataDatasets_ShouldReturnFalse()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            var tag1 = DicomTag.URNCodeValue;
-            var tag2 = DicomTag.SelectorLTValue;
-            dataset1.Add<string>(tag1);
-            dataset2.Add<string>(tag2);
-
-            Assert.False(dataset1 == dataset2);
-        }
-
-        [Fact]
-        public void InequalityOperator_DifferentData_ShouldReturnTrue()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            var tag1 = DicomTag.URNCodeValue;
-            var tag2 = DicomTag.SelectorLTValue;
-            dataset1.Add<string>(tag1);
-            dataset2.Add<string>(tag2);
-
-            Assert.True(dataset1 != dataset2);
-        }
-
-        [Fact]
-        public void InequalityOperator_SameDataDatasets_ShouldReturnFalse()
-        {
-            var dataset1 = new DicomDataset();
-            var dataset2 = new DicomDataset();
-            var tag = DicomTag.URNCodeValue;
-            dataset1.Add<string>(tag);
-            dataset2.Add<string>(tag);
-
-            Assert.False(dataset1 != dataset2);
-        }
-
         #endregion
 
         #region Support methods
